requirements.txt
Django==3.2
djangorestframework==3.12
gunicorn==20.1
celery==5.0
redis==3.5